test(issue-agent): keep review context bounded
