+++ docs/agents/issue-agent.md
@@ -64,11 +64,9 @@ unbounded repair attempts.
 An Issue by an `OWNER`, `MEMBER`, or `COLLABORATOR` with current `write`,
 `maintain`, or `admin` permission may start automatically. Other reports need
 an exact first-line `/agent fix` from an actor whose permission is re-read.
-For a trusted Issue author, a failed or temporarily non-write permission read
-is retried inside one bounded 3.1-second consistency window. A persistent API
-failure aborts the Controller without recording the author as unauthorized; a
-stable non-write result still waits for authorization. Issue association never
-substitutes for current permission.
+For a trusted Issue author, failed or temporarily non-write permission reads
+are retried for at most 3.1 seconds. Only a fresh `write`, `maintain`, or
+`admin` result authorizes admission; API errors abort and non-write results wait.
 The other commands are:
 
 - `/agent retry` — one fresh ephemeral attempt after `needs_human`;
